+++ Sources/App/Models/Sale.swift
@@ -21,7 +21,7 @@ final class Sale: Model, Content {
     @Field(key: "saleDate")
     var saleDate: Date
     @Field(key: "total")
-    var total: Double
+    var total: Int
     
     @Timestamp(key: "created_at", on: .create)
     var createdAt: Date?
@@ -43,7 +43,16 @@ final class Sale: Model, Content {
     
     init() { }
     
-    init(id: UUID? = nil, paid: Bool, paymentType: String, saleDate: Date, total: Double, subsidiaryID: Subsidiary.IDValue, customerID: Customer.IDValue?, employeeID: Employee.IDValue) {
+    init(
+        id: UUID? = nil,
+        paid: Bool,
+        paymentType: String,
+        saleDate: Date,
+        total: Int,
+        subsidiaryID: Subsidiary.IDValue,
+        customerID: Customer.IDValue?,
+        employeeID: Employee.IDValue
+    ) {
         self.id = id
         self.paid = paid
         self.paymentType = paymentType
